Update HH-of-T3.sage
I believe there was a mistake when counting the dimensions (or equivalently the order of the groups). It used to calculate the sum of the invariants, but it should calculate the product. e.g. for Z/2 x Z/2 x Z/2 the order is 2*2*2 = 8, not 2+2+2=6.

Additionally, before we needed to add the max(1,...), for the case when we had the trivial group, which should count as 1 and not as 0. We can drop the max function, cause product(empty) gives already 1, as need be. 

Greetz Mat